requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
pytest
httpx

